requirements.txt
pydantic==2.8.2
requests==2.32.3
supabase==2.7.3
typing-extensions==4.12.2
PyYAML==6.0.2
Jinja2==3.1.4